feat: move filters into a sub-package
These really don't belong in the root package.
